Add GCC attributes to printf functions.
While doing the order1 code I noticed our stdio library is missing
some of the automatic GCC error checking from the system's
library.  I added that back in and fixed a bunch of cases in the
code were we would crash where trying to print error messages.

bk: 4d34cdeesE_CO43m51sirj2ZMds7qg

+++ src/bkd_http.c
@@ -806,7 +806,7 @@ http_anno(char *page)
 			p = src;
 			while (*p && (isalnum(*p) || (*p == '_'))) p++;
 			if (*p == '(') {
-				printf("<a name=\"%.*s\">", p-src, src);
+				printf("<a name=\"%.*s\">", (int)(p-src), src);
 				closeTag=1;
 			}
 		}
